pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qa-automation-template"
version = "0.1.0"
description = "General-purpose pytest automation framework template for web, API, and mobile"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pytest>=8.0.0",
  "pytest-xdist>=3.6.0",
  "allure-pytest>=2.13.0",
  "requests>=2.31.0",
  "selenium>=4.18.0",
]

[project.optional-dependencies]
windows = [
  "Appium-Python-Client>=3.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
